+++ src/backend/joanie/core/filters/admin/__init__.py
@@ -333,6 +333,59 @@ class OrderAdminFilterSet(filters.FilterSet):
         ).distinct()
 
 
+class BatchOrderAdminFilterSet(filters.FilterSet):
+    """
+    FilterSet to filter batch orders by several criteria.
+    """
+
+    class Meta:
+        model = models.BatchOrder
+        fields: List[str] = ["query"]
+
+    query = filters.CharFilter(method="filter_by_query")
+    state = filters.ChoiceFilter(choices=enums.BATCH_ORDER_STATE_CHOICES)
+    payment_method = filters.ChoiceFilter(
+        choices=enums.BATCH_ORDER_PAYMENT_METHOD_CHOICES
+    )
+    organization_ids = filters.ModelMultipleChoiceFilter(
+        queryset=models.Organization.objects.all().only("pk"),
+        field_name="organization",
+        distinct=True,
+    )
+    product_type = filters.MultipleChoiceFilter(
+        field_name="relation__product__type",
+        choices=enums.PRODUCT_TYPE_CHOICES,
+    )
+    product_ids = filters.ModelMultipleChoiceFilter(
+        queryset=models.Product.objects.all().only("pk"),
+        field_name="relation__product",
+        distinct=True,
+    )
+    course_ids = filters.ModelMultipleChoiceFilter(
+        queryset=models.Course.objects.all().only("pk"),
+        field_name="relation__course",
+        distinct=True,
+    )
+    ids = MultipleValueFilter(field_class=fields.UUIDField, field_name="id")
+
+    def filter_by_query(self, queryset, _name, value):
+        """
+        Filter resource by looking for product title, course title | code, organization
+        title | code, owner username, email, full_name
+        """
+
+        return queryset.filter(
+            Q(relation__course__translations__title__icontains=value)
+            | Q(relation__course__code__icontains=value)
+            | Q(organization__translations__title__icontains=value)
+            | Q(organization__code__icontains=value)
+            | Q(relation__product__translations__title__icontains=value)
+            | Q(owner__email__icontains=value)
+            | Q(owner__username__icontains=value)
+            | Q(company_name__icontains=value)
+        ).distinct()
+
+
 class DiscountAdminFilterSet(filters.FilterSet):
     """
     DiscountAdminFilterSet allows to filter this resource with a query for amount and rate.